DOR-383 Enable warnings as errors in dorado code, and disable warnings from edlib and minimap2.  Fixed up current warnings.

+++ dorado/cli/duplex.cpp
@@ -207,7 +207,7 @@ int duplex(int argc, char* argv[]) {
             converted_reads_sink = aligner;
         }
         auto read_converter = pipeline_desc.add_node<ReadToBamType>(
-                {converted_reads_sink}, emit_moves, 2, 0, nullptr, 1000);
+                {converted_reads_sink}, emit_moves, 2, 0.0f, nullptr, 1000);
         auto duplex_read_tagger = pipeline_desc.add_node<DuplexReadTaggingNode>({read_converter});
         // The minimum sequence length is set to 5 to avoid issues with duplex node printing very short sequences for mismatched pairs.
         std::unordered_set<std::string> read_ids_to_filter;
@@ -216,7 +216,7 @@ int duplex(int argc, char* argv[]) {
                 read_ids_to_filter, 5);
 
         std::vector<dorado::stats::StatsCallable> stats_callables;
-        ProgressTracker tracker(num_reads, duplex);
+        ProgressTracker tracker(int(num_reads), duplex);
         stats_callables.push_back(
                 [&tracker](const stats::NamedStats& stats) { tracker.update_progress_bar(stats); });
         stats::NamedStats final_stats;
@@ -357,7 +357,7 @@ int duplex(int argc, char* argv[]) {
             }
             pipelines::create_stereo_duplex_pipeline(
                     pipeline_desc, std::move(runners), std::move(stereo_runners), overlap,
-                    mean_qscore_start_pos, num_devices * 2, num_devices,
+                    mean_qscore_start_pos, int(num_devices * 2), int(num_devices),
                     std::move(pairing_parameters), read_filter_node);
 
             pipeline = Pipeline::create(std::move(pipeline_desc), &stats_reporters);
